Add checkstyle to persistence module
