+++ common/persistence/src/main/java/app/coronawarn/server/common/persistence/domain/DiagnosisKeyBuilder.java
@@ -1,5 +1,11 @@
 package app.coronawarn.server.common.persistence.domain;
 
+import static app.coronawarn.server.common.persistence.domain.DiagnosisKeyBuilders.Builder;
+import static app.coronawarn.server.common.persistence.domain.DiagnosisKeyBuilders.FinalBuilder;
+import static app.coronawarn.server.common.persistence.domain.DiagnosisKeyBuilders.RollingPeriodBuilder;
+import static app.coronawarn.server.common.persistence.domain.DiagnosisKeyBuilders.RollingStartNumberBuilder;
+import static app.coronawarn.server.common.persistence.domain.DiagnosisKeyBuilders.TransmissionRiskLevelBuilder;
+
 import app.coronawarn.server.common.protocols.external.exposurenotification.Key;
 
 /**
@@ -50,57 +56,4 @@ public class DiagnosisKeyBuilder implements Builder, RollingStartNumberBuilder,
     return new DiagnosisKey(
         this.keyData, this.rollingStartNumber, this.rollingPeriod, this.transmissionRiskLevel);
   }
-}
-
-interface Builder {
-
-  /**
-   * @param keyData generated diagnosis key.
-   * @return this Builder instance.
-   */
-  RollingStartNumberBuilder withKeyData(byte[] keyData);
-
-  /**
-   * @param protoBufObject ProtocolBuffer object associated with the temporary exposure key.
-   * @return this Builder instance.
-   */
-  FinalBuilder fromProtoBuf(Key protoBufObject);
-}
-
-interface RollingStartNumberBuilder {
-
-  /**
-   * @param rollingStartNumber number describing when a key starts. It is equal to
-   *                           startTimeOfKeySinceEpochInSecs / (60 * 10).
-   * @return this Builder instance.
-   */
-  RollingPeriodBuilder withRollingStartNumber(long rollingStartNumber);
-}
-
-interface RollingPeriodBuilder {
-
-  /**
-   * @param rollingPeriod Number describing how long a key is valid. It is expressed in increments
-   *                      of 10 minutes (e.g. 144 for 24 hours).
-   * @return this Builder instance.
-   */
-  TransmissionRiskLevelBuilder withRollingPeriod(long rollingPeriod);
-}
-
-interface TransmissionRiskLevelBuilder {
-
-  /**
-   * @param transmissionRiskLevel risk of transmission associated with the person this key came
-   *                              from.
-   * @return this Builder instance.
-   */
-  FinalBuilder withTransmissionRiskLevel(int transmissionRiskLevel);
-}
-
-interface FinalBuilder {
-
-  /**
-   * @return {@link DiagnosisKey} instance
-   */
-  DiagnosisKey build();
 }
